### 2018-06-08 #222 @maelle
Thanks a lot @hrbrmstr!

Thanks for your submission @leonawicz, I'll proceed with other checks now that overlap with `pubcrawl` is clarified.
