### 2018-06-08 #222 @hrbrmstr
Email & I have been bitter, bitter enemies of late. So it's more of a "wall" between me & it vs bits slipping through cracks :-)

I had an inkling that the style sheet transformer might be super aggressive for an epub (and epubs have, er, "diverse" XML formats, too) that your straightforward approach is likely much, much safer.

I was going to suggest using `https://github.com/hrbrmstr/freebase` (since it'd cover the `purrr` and some `dplyr` use cases in the pkg pretty well) but you'd still need `tidyr`. Plus, it's super likely that a majority of the pkg audience will be tidyverse users.

If it's OK with y'all I'll likely set the `pubcrawl` repo to an "archive" state and replace the README with a link to the rOpenSci'd version of `epubr` once the review is done. If so, pls @ me on the final review thread box and I'll get'er done.

Again, rly nice work!
