### 2018-06-08 #222 @leonawicz
Thanks! :) I'll have to go back and look at the details when I have a chance, but I can probably remove `purrr` in a later package version by replacing with some uglier code, but I think `dplyr` is the heavier dependency, and unfortunately I the one I particularly want to retain.

The diversity of formats is also why I could not be as aggressive as I would have liked with things like the optional attempts at chapter identification. I was bummed to see that of all epubs I looked at, go figure, the example Project Gutenberg book I included as an example file in the package has some of the most problematic formatting. That bummed me out haha. I'm guessing it's probably really common, but so far it's the only book I've seen where the internal html files that mark the different book sections do can actually begin and end in the middle of chapters, as though they cared more about keeping each html file approximately the same size rather than breaking the text into meaningful sections. So lame! :)

In a future version I would like to include a function that loads a book in the browser as a simple e-reader using `shiny` perhaps, but after seeing the arbitrary section breaks in that book I had to put that thought on hold.

Thanks again for your help!
