### 2018-06-08 #222 @maelle
Nice to see such a constructive discussion even before the actual review process starts 😀 Thanks again for chiming in @hrbrmstr

Btw my checks are still pending but I'll probably recommend to add a link to https://github.com/ropenscilabs/gutenbergr in the README @leonawicz 😉
