### 2018-06-08 #222 @leonawicz
@maelle Done, via latest commit. Thank you for the suggestion. It's under the Reference heading. Let me know if I should format it differently.
