### 2018-06-09 #222 @maelle
Thanks again for your submission @leonawicz!

### Editor checks:

---

#### Editor comments

The README states "Subsequent versions will include a more robust set of features.". Before I start looking for reviewers, could you

* List what these features would be (e.g. as issues in your repo associated to a [milestone](https://help.github.com/articles/about-milestones/) )

* Confirm you don't want to implement any of them before or during the review process.

We can put this submission on hold so that the reviewers look at a more finished product, which in my opinion would help you make the most of our onboarding process.

In any case, writing down what your ideas are will help lessen duplication of efforts, since reviewers often suggest enhancements.
---

Reviewers:
Due date:
