### 2018-06-11 #222 @leonawicz
Badge added. Thank you.
