### 2018-07-01 #222 @suzanbaert
Yes! I already did some experimentation and i have time set apart on monday to finish the last few items. But first a beautiful summer sunday 😉
