### 2018-07-02 #222 @suzanbaert
>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

First of all, I love the idea of the package and the ability to add a vector of epubs. I also was quite amazed at the speed, I would have thought it was going to take longer to turn a few quite sizeable books into an R object. Functions also work as part of a pipe, and plays very nice with dplyr.

A few suggestions though:

**Error messages**
- When adding a non valid file (file doesn't exist or was wrongly spelled) the error message is quite mysterious: `Error in if (grepl("<|>", x)) { : argument is of length zero In addition: Warning message: In utils::unzip(file, exdir = exdir) : error 1 in extracting from zip file.` It might be worth adding a friendlier error message that just reads "file could not be found" or "file is not a valid epub file" or something similar?

- Similarly, it seems that title is a mandatory field in the fields argument. When 'title' is not present, you get an error: `Error: Column 'title' must be a 1d atomic vector or a list. In addition: Warning message: Unknown or uninitialised column: 'title'.` I would suggest to either make 'title' a standard field that automatically gets added (like data) or to make title unnecessary.

**Function naming**
Secondly, regarding the function names, `epub_meta` and `epub_unzip` are perfectly in line with the rOpenSci's packaging guidelines, but the other two are not. I was wondering especially about `epub`, it would intuitively make more sense to name it `read_epub()` which would make it more in line with nearly all other import functions in R people will know about. Alternative could be `epub_import` or something similar?

Similarly with `first_nchar`, perhaps something like `epub_glimpse()` or anything similar that does start with epub_ and makes it clear you're getting a sneak peek on the file?

**README documentation**
Last but not least, though everything is in the documentation and R veterans probably will have no issues, I would make the documentation a bit more beginner's friendly.
One element in particular caught my eye: the instructrions in the readme to import an epub. It's probably silly, but when I first installed and tried to import an epub (i was in play-modus and so not very concentrated), but I also did a system.call() followed by epub(). I somehow thought I had to make a system file from my epub and then sort of unzip it. (Needless to say that I never used the system.call() function before so wasn't entirely sure what it did...)

It took only a mysterious error and a bit more focus to realize my silly mistake, but I did wonder whether the README could just be a bit clearer. Perhaps something along these lines? Tiny change, but maybe a few frowns less from people first using the package?

```
#importing an epub file
x <- epub("path/to/epubfile.epub")

#to use the example dracula.epub that is built-in with epubr:
file <- system.file("dracula.epub", package = "epubr")
x <- epub(file)
```

<br><hr>
Output of devtools checks:

**Devtools checks**

```
# results devtools::test()
test-lintr.R:4: warning: Package Style
incomplete final line found on 'E:/Rprojects/epubr/.lintr'

test-lintr.R:4: error: Package Style
Invalid DCF format.
Regular lines must have a tag.
Offending lines start with:
  inst/.lintr
1: lintr::expect_lint_free() at E:\Rprojects\epubr/tests/testthat/test-lintr.R:4
2: lint_package(...)
3: read_settings(path)
4: read.dcf(config_file, all = TRUE)
5: stop(gettextf("Invalid DCF format.\nRegular lines must have a tag.\nOffending lines start with:\n%s",
       paste0("  ", lines, collapse = "\n")), domain = NA)
```

Regarding `devtools::check()`, it complained about two packages i do not have installed on this PC, but there also was a warning regarding `qpdf`.

```
# results devtools::check()

Status: 1 WARNING, 1 NOTE

* checking package dependencies ... NOTE
Packages suggested but not available for checking: 'lintr' 'covr'

* checking for unstated dependencies in examples ... OK
 WARNING
'qpdf' is needed for checks on size reduction of PDFs
```
