### 2018-07-02 #222 @calebasaraba
6-8

---

I think this is a solid idea for a package and it accomplishes its main goal well: to provide users with functions to transform `epub` formatted ebooks into neatly nested R dataframes that can be used for further textual analyses. Very well done!

I found that the package was easy to use and performed well at these actions using R 3.5 on windows and mac systems. I wrote down some of my thoughts below:

__Function commenting__
While working through the code in `epub.R`, I had a bit of a hard time parsing the multitude of `if` statements. I understand that they are necessary, but I was wondering if there was a stylistic way to structure them in the code to be easier for a new user sifting through the function details. Perhaps one way would be to change some of the single-line if statements into curly-braced statements with indentations, or providing a few more in-line comments to explain what's happening. I mention this because, due to the heterogeneity of `epub` formats, I think a lot of your more adventurous users will likely be looking at your actual code while trying to learn more about `epub` processing and how they can tweak it to their own unique situation / format.

__Function naming__
It may make more sense to change the `epub()` function name to the `object_verb` style recommended by the ROpenSci packaging guidelines, something like `epub_import()` or switching its name with the internal `epub_read()`. Similarly, `first_nchar()` could be something like `epub_head()`.

__Documentation note__
I understand from your post above that you're trying to limit dependencies, and you've done an admirable job. I'm not sure if you avoided it for a particular reason, but I've found it easier to use `@importFrom` in `roxygen2` when documenting a particular function rather than remembering what packages I'm using throughout and adding them to the `DESCRIPTION` manually, which I think is how you did it here?

__Chapter separation functionality__
I didn't have any problems using the `epub()` or `epub_unzip()` functions on either mac or windows using R version 3.5.

I downloaded several epub files from project gutenberg and played around with loading them in with the `epubr` package. In each instance, the metadata fields were easily readable and correctly formatted, but I could never really get the chapter separation to work very well, and I did not find the default item separation useful / meaningful. I understand how challenging it would be to get this sort of separation functionality to work across a wide variety of `epub` files, and you've put in effort to make it work at all -- but I wonder if it philosophically makes sense to make chapter and/or item separation the default behavior if it is not very generalizable (or in the case of item, that useful?). Perhaps it would be better have the default nested dataframe structured more like the output from `gutenbergr`, which seems to be a straightforward reproduction of the text present in the `epub` file?

__Documentation / Vignettes__
I like that you chose to include an `epub` file (Dracula) that was poorly formatted for demonstration purposes in your vignette, but it also might be instructive for your users to show an example where the chapter separation works smoothly, if possible (this would be especially true if you change the default chapter/item separation behavior and want to highlight this as optional functionality).

On the topic of chapter separation / funky formatting, you include the following paragraph that refers to some sort-of hidden functionality:

```
These developmental arguments are currently undocumented,
though they can be explored if you are inclined to read the
package source code and pass additional arguments to ....
They have been tested successfully on many e-books, but
certainly not a representative sample of all e-books.
The approaches these arguments use may also change before
they are formally supported and explicitly added to a future
version of the package.
```
I wonder if it is a necessary addition to the vignette, since without documentation or examples of test cases, it only serves to tantalize the reader! I noticed earlier in the submission issue you mentioned that you didn't have any concrete next steps for development, so perhaps some examples or explanation of these arguments could be one; alternatively, you could consider removing this reference from your documentation.

---
