### 2018-07-03 #222 @maelle
Thanks a lot for your review @calebasaraba! 🥅 ⚽️ 😃

@leonawicz now both reviews are in! As per our [author's guide](https://ropensci.github.io/dev_guide/onboarding-guide-for-authors.html) we ask that you respond to reviewers’ comments within 2 weeks of the last-submitted review, but you may make updates to your package or respond at any time.

Two notes:

* Reg imports I don't know what @calebasaraba meant since I haven't looked at the code as thoroughlyt as he did but as a reminder we recommend using the pkg::function syntax. Cf http://r-pkgs.had.co.nz/namespace.html#imports It doesn't add packages to DESCRIPTION though. `usethis::use_package` allows one to add packages to DESCRIPTION more automatically, and [this new package of Hadley Wickham's](https://github.com/hadley/requirements) might help finding all used dependencies.

* If you rename the functions have a look at [our guidance regarding "changing stuff in your package"](https://ropensci.github.io/dev_guide/evolution.html)
