### 2018-07-03 #222 @leonawicz
Hi all! Thank you all so much for your feedback and hard work! It is much appreciated. :) Some changes have been implemented already. Those and other replies below.

#### Re: cryptic error message.
Agreed, thank you for catching that. I have made more informative error messages for missing files and non-epub files, for epub, epub_meta, and epub_unzip.
#### Re: required title field.
This was a big issue. Thank you again. I have made `title` automatic like `data`.

Context: title is important because it acts as the most likely/dependable field to uniquely identify books when `file` is a vector.

Of course, multiple books could have the same title, such as multiple editions of one book. Title is not guaranteed to be unique. However, it is the best bet. As such, I have made `title` the field that falls back on file names (mostly likely unique identifier) when `title` is not a field in the metadata. `title` can also be passed as an additional argument (`...`) to `epub` so the user can override the name of the metadata entry containing title information. See updated help doc for details.

I have not encountered any epubs yet lacking a title field, but I did my best to test this out by passing the additional `title` argument using non-existing field names.
#### Re: function naming.
I am open to renaming `first_nchar` to something like `epub_text`, `epub_head`, `epub_glimpse`.
I suppose given the way it works, I am partial to `epub_head`. There could also be a tail function (though probably not much use), or a custom slice (could be useful), in addition to just looking at the first n characters. I think this makes `*head` the best suffix for the current function.

I'm not inclined to lengthen `epub` with an additional verb. In my view, `epub` is meaningful, clear, but also appropriate for such a small package with such a singular use. It's easier to view it as the master, over-arching package function. It also does not live alongside the extraneous functions like `epub_meta` and `epub_unzip` but wraps around them and encapsulates the whole package purpose. I don't think there is really anything else in such a small package to confuse `epub`'s use with.

#### Re: documentation.
The example using `system.file` is pretty standard, but I have added a comment above the line to make it more clear what it means.
#### Re: vignette.
I agree I probably did more harm than good by indicating that there is developmental functionality potentially accessible via `...` that I want to integrate explicitly later but which remains currently far too open to changing approaches at this time. I have removed the reference from the vignette.
#### Re: dependencies/import note
I'm confused regarding this comment, sorry. I used `usethis` and explicit namespaces in the package. I used `@importFrom` only if necessary, e.g. `@importFrom magrittr %>%`.
#### Re: chapter separation
I'm unclear about "did not find the default item separation useful / meaningful". Because this is generally unpredictable, the default is to attempt no chapter identification. `drop_sections` and `chapter_pattern` are both `NULL` unless the user says otherwise in the context of their known e-books.

The default is to simply dump each book section into its own row of the `data` data frame (you may mean this). My approach is to leave this structure as is, and users can alter it further however they want. For many books, this is a very good format because this does give a meaningful and useful breakout of sections for many books. It also is in keeping with the general approach in `epubr` of not altering the file content (e.g., by automatic text cleaning steps that users can't prevent) other than to basically strip the xml tags. Things are made human readable but otherwise left close to the original table structure.

### Related suggestion and some context:

#### It is very tempting to us, but Project Gutenberg maybe is getting too much attention :)

I think we should not pay much attention to Project Gutenberg (PG) texts honestly. It is a convenient place to grab a public domain epub file for package testing and ability to share the content/examples.
In fact, this is why I included one in the package. However, my perspective is that PG is potentially highly unrepresentative of the e-books users will tend to load with `epubr` as well as not a very realistic use case:

I have looked at books from a variety of publishers. I don't consider this to be representative of "all epubs" of course, but I did notice that none had formatting like PG books I've looked at (which I admit is far fewer). I have looked at a lot of licensed/copyrighted books, the type you have to purchase, from a range of publishers and years. I consider these much more representative of the types of e-books users will load with `epubr`. This was the motivation for the package.

Much of the public domain stuff can be obtained in multiple formats and pulled directly into R from the web. `epubr` is better for when you have local files, and likely nothing but the local files in this single format. For licensed/purchased books, users probably only have one copy in one format. It may not even be epub originally, but they may have locally converted it to epub (say, from .mobi after buying from Amazon, or something similar) using other software like `Calibre`. This is more fundamentally what `epubr` is for.

With the goal of loading a novel's text into R to be used in a subsequent text analysis, when it comes to PG there is no reason at all to use `epubr`. To do so requires creating an unnecessary intermediary step of obtaining local epub format files from PG, then loading them into R with `epubr`, but none of that is necessary. All of that can easily be skipped, for example by just using `gutenbergr` if I'm not mistaken.

So overall, I think PG is not only not a good/gold standard for evaluation, but actually an exceptional edge case/not a representative use case for `epubr` or of epubs in general. Second, PG also has no need at all for an `epubr` or other type of intermediary `epub` format processing stage. It's hard to say what *is* most representative of "all books", or at least "all popular books," but I am pretty confident that PG is not it. I would say `epubr` is more for use cases where the files themselves, the users' collections of local epub files, represent the bottleneck of their entire analysis, in that it may be the sole source of the particular data that a user has access to. Apart from specific research involving PG titles, the intersection of most users' e-book collections with the public domain works on PG is probably very small.

Thoughts?

#### Re: Devtools checks
test-lintr.R:4: warning: Package Style
incomplete final line found on 'E:/Rprojects/epubr/.lintr'

I think this can be ignored. It is possible that if you pull the repo down to some systems, the symlink stops being seen as a symlink (Windows?). On my local Windows it is fine (no warnings). Recreating the symlink can fix the problem on a local system. But this passes on Windows, Mac and Linux (Travis-CI, I think with warnings as errors). This file just links to the real lintr file in the `inst` directory. This setup may look weird but it has to do with getting lintr to work with unit testing on systems like Travis-CI for package builds.

test-lintr.R:4: error: Package Style
Invalid DCF format.
Regular lines must have a tag.
Offending lines start with:
  inst/.lintr
1: lintr::expect_lint_free() at E:\Rprojects\epubr/tests/testthat/test-lintr.R:4
2: lint_package(...)
3: read_settings(path)
4: read.dcf(config_file, all = TRUE)
5: stop(gettextf("Invalid DCF format.\nRegular lines must have a tag.\nOffending lines start with:\n%s",
       paste0("  ", lines, collapse = "\n")), domain = NA)

I'm not sure if this is more of the same. I am unable to reproduce these errors locally or remotely.

"warning regarding qpdf."

It's been a long time since I installed qpdf, so I don't remember the exact complaint, but if I *didn't* have it installed then local package builds threw some kind of message or warning about it not being available.

Status: 1 WARNING, 1 NOTE
* checking package dependencies ... NOTE
Packages suggested but not available for checking: 'lintr' 'covr'

I think this is normal, just need to have them on the system if you are building a package that uses them. I'm saying "I think" a lot, which is just to say, if anything I'm saying is not quite right, please feel free to jump in and set things on the right track.

* checking for unstated dependencies in examples ... OK
 WARNING
'qpdf' is needed for checks on size reduction of PDFs

 Yeah, I think this is the warning I get rid of by making sure qpdf is installed.

#### Re: function commenting/code
This is one area where it's possible it may have been premature for me to submit my package for review?  I can't speak to what other downloaders of `epubr` are doing with it, but just myself I have plenty of unofficial use cases already in other projects that are dependent on the still-evolving/undocumented code inside some of the package functions. I can't remove it, but I also can't really say what it will look like later (or when I will get to it, since this is a personal project).

I published to CRAN because what was official/documented, was ready to go. The parts that are in the works though, it's just too early to say exactly what that will look like and trying to document it or give examples would be getting into the weeds I think. Please let me know what you think. I can always resubmit in the future if there is too much problematic about the state of some of the internals.

Thanks again so much! :) Please forgive me if I forgot to address some comments. Or if I shouldn't have addressed both reviews in one reply, but I saw some overlap. It took a while to compile this (probably not nearly as long as it took you to review the package code and I am very grateful). Please let me know if I missed anything or confused anything.
