### 2018-07-09 #222 @maelle
Btw maybe relevant https://github.com/hadley/ellipsis
