### 2018-07-11 #222 @leonawicz
Thanks. I have updated the documentation to indicate that, with the exception of the new optional `title` argument override that is now accepted by `...` (and documented in the help Details), `...` is currently developmental/unsupported.

I have also renamed `first_nchar` to `epub_head`. I began to make the alias and add the relevant warning to `first_nchar`, but then stopped and decided to simply rename the function because I realized `first_nchar` had never been released. It was just added at the time of this review. It's not in the CRAN release or any official GitHub release even.
