### 2018-07-11 #222 @suzanbaert
Hi Matt,

I've downloaded the latest version from GitHub, and tried with latest changes.

Re: Error messages.
Much clearer! Thanks for fixing that.

Re: Title automatically added.
Again, thanks for fixing, it was a minor comment, but it just felt weird to not have flexibility when you thought you'd had some. I think this is clearer, thanks.

Re: epub_head.
I like the new function name
Regarding the epub() or read_epub() - as @calebasaraba said, it's really up to you. I have no issue with it being the former given the scope of the package, it was just stylistic feedback because it's nice as a user when similar functions have same syntax irregardless of which package they are coming from. No showstopper whatsoever, it's just something I thought about.

Re: types of epub
I actually tried about 5 other epubs that were not Gutenberg, and I had no issues. The text even came out clearer than the Gutenberg - I guess because project Gutenberg adds these little lines.
The only one I tried that really came out impossibly messy was a technical epub book on powershell with a lot of code blocks. That one was destined to be really messy and not really due to the package :wink:

The devtools notes and warnings did not seem to have any impact on the package properly working, so I only passed those on in case it was useful to you.
