### 2018-07-11 #222 @leonawicz
Thanks Suzan!

Yeah, I also tried a number of Star Trek books (related to my `rtrek` package) using `epubr`. Most were novels, but a few were more like manuals or reference books. They don't work *as* well (and some don't have what would meaningfully be called "chapters"), but I took nominal steps toward making books like those still somewhat parsable. `epubr` is definitely more interesting for popular literature. But that's just my personal opinion/biased interest.
