### 2018-07-11 #222 @suzanbaert
I agree! I think people will use function, not so much technical literature. I wasn't expecting the powershell one to come out good, but i was curious :)
