### 2018-07-11 #222 @leonawicz
Also just for more insights, wanted to share, I did an interesting test.

Where one book had not only chapters, marked by `ch\\d\\d`, but also a "Part" 1, 2 and 3 dividing up the chapters into three groups. Those were marked in the metadata as `part\\d`. Each "Part" section had literature, just like the chapters did. Additionally, there were intro, epilogue, and a couple other sections that still had the "real" book text in them.

I read the book with `epub` several times, each time adding a more inclusive regex pattern, e.g., using `"ch\\d\\d"`, then using `ch\\d\\d|part\\d`, and so on, which demonstrates how easy it is to use the `chapter_pattern` argument to define what you want to distinguish from the rest of the book as "chapters". In this case, someone could decide they strictly want to single out the actual chapters, or they could fold in the part breaks and prologue/epilogue, since those technically include part of the actual story.

It's difficult to show these nice examples when books are licensed works that can't be distributed.

Also, one thing I shared with @maelle on Slack that I think might not have been mentioned elsewhere is that when you do perform chapter identification on books using a regex pattern, `epubr` is, in general, able to reorder the sections appropriately by chapter sequence. Many books I have seen, even when they have really nice metadata formatting, have their book sections occur in an arbitrary order, giving a scrambled appearance to the resulting `data` rows. So `epubr` can handle this elegantly is a lot of cases.

Other text parsing packages, like the more general `tika` package I saw recently, just dump contents into a single character string, and in such cases I found that the text of course was all out of order. `epubr`'s output structure allows text order to be maintained if important to the user, and chronological order can actually be established even when reading the raw file would otherwise load text in a more scrambled order.
