### 2018-07-16 #222 @maelle
Approved! 🚀  Thanks @leonawicz  for submitting and thanks @calebasaraba @suzanbaert  for your reviews! 😺

To-dos:
"```
[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)```"
- [ ] Make a PR to `pubcrawl` with a link to the repo (under the ropensci organization) since @hrbrmstr offered to archive his repo [in this comment](https://github.com/ropensci/onboarding/issues/222#issuecomment-395788364).

Should you want to awknowledge your reviewers in your package DESCRIPTION, you can do so by making them `"rev"`-type contributors in the `Authors@R` field (with their consent).  More info on this [here](https://ropensci.github.io/dev_guide/building.html#authorship).

Welcome aboard! We'd also love a blog post about your package, either a short-form intro to it (https://ropensci.org/tech-notes/) or long-form post with more narrative about its development. (https://ropensci.org/blog/). If you are interested, @stefaniebutland will be in touch about content and timing.
