### 2018-07-17 #222 @leonawicz
Thank you! @maelle I've added the footer and the related packages section. I used the `@` syntax for authors you suggested, thinking that when it appear on GitHub it was intended to act as a link to each author's profile or something. But maybe you did not mean that. I'm thinking I should switch it to people's proper names then? There is already a link to each package mentioned.

I added the `codemeta`.

I'll do the transfer soon.

I'm not sure what the pull request to `pubcrawl` should entail. I have not done anything like this before regarding archiving packages. Also, aside from the AppVeyor line mentioned above, are the other substitutions as simple as swapping out my account name for `ropensci`? I will also update the url in the package hex sticker logo.

Thanks! 😸
