### 2018-07-18 #222 @maelle
Thanks!

* No for authors it means using the Authors@R field in DESCRIPTION, sorry it was unclear.

* Update the thread once the transfer is done. Yes it should actually be that simple. There are two URLs to update in DESCRIPTION too (URL and BugReports fields).

* Ok will do the PR and tag you!
