### 2018-07-19 #222 @maelle
The blog post discussion can continue, but closing this since all transfer steps have now been done!
