### 2018-11-18 #222 @stefaniebutland
@leonawicz In light of your tweet https://twitter.com/leonawicz/status/1064224408900820992 we'd welcome a technote on this update if you're interested. (thanks for the heads up @maelle)
https://ropensci.org/technotes/
